Add files via upload
The code here is to create various Data Types, their methods and properties.
The properties explored includes but not limited to the following:
i. Use of Arrays
ii. Use of Strings
iii. Concatinating Strings
iv. Calculations using Javascript and adding of items to a list using Unshift and push commands.